requirements.txt
Flask
tensorflow
numpy
Pillow
pyngrok
